test fixes

+++ handler/oauth2/flow_client_credentials_test.go
@@ -76,7 +76,7 @@ func TestClientCredentials_HandleTokenEndpointRequest(t *testing.T) {
 					Audience:             []string{"https://www.ory.sh/api"},
 					AccessTokenExpiresAt: 3600000000000,
 				}).MaxTimes(2)
-				client.EXPECT().GetAccessTokenExpiresAt().Return(3600000000000)
+				client.EXPECT().GetAccessTokenExpiresAt().Return(int64(3600000000000))
 			},
 		},
 		{
@@ -90,7 +90,7 @@ func TestClientCredentials_HandleTokenEndpointRequest(t *testing.T) {
 					Scopes:               []string{"foo"},
 					AccessTokenExpiresAt: 3600000000000,
 				}).MaxTimes(2)
-				client.EXPECT().GetAccessTokenExpiresAt().Return(3600000000000)
+				client.EXPECT().GetAccessTokenExpiresAt().Return(int64(3600000000000))
 			},
 		},
 		{
@@ -105,7 +105,7 @@ func TestClientCredentials_HandleTokenEndpointRequest(t *testing.T) {
 					Scopes:               []string{"foo", "bar", "baz"},
 					AccessTokenExpiresAt: 3600000000000,
 				}).MaxTimes(2)
-				client.EXPECT().GetAccessTokenExpiresAt().Return(3600000000000)
+				client.EXPECT().GetAccessTokenExpiresAt().Return(int64(3600000000000))
 			},
 		},
 	} {
